Using creation status for sharepoint info
Previously we had been grabbing the data about the new sharepoint after it had been created. This used to provide information about the new sharepoint within seconds. it now can take up to 30 min before the sharepoint can be queried. Instead we are now pulling the information provided about the sharepoint group at the time of creation and then making a summary with that data.